Windows Update Repair Utility
A batch file that performs a multi step process to help fix any Windows Update Issues. Each step is bypassable in case a user already performed the action previously. The actions performed by the batch file are as follows:

1. Open the Windows Update Troubleshooter
2. Restart Windows Update Related Services (bits, wuauserv, appidsvc, cryptsvc)
3. Run sfc/scannow and DISM
4. Restart the pc or exit the program

Note: This batch file must be run as admin due to sfc and DISM commands. If the user does not run as admin an error message will be displayed asking the user to rerun the script with elevated privileges.