Automated Battery Report Generation
A batch file that automates the generation, saving and opening of a battery health report for Windows laptops. The goal of this script is take the battery report command that already exists in Windows, and make it more seamless and less tedious to use. Here is an overview of the script does:

1. Generates a battery report by running the CMD command 'pwrcfg /batteryreport' and saves the report to the desktop rather than the default 'C' directory.
2. Automatically opens the report in the default browser.

This script makes using the battery report command less tedious to use by saving the report in an easy to find location and eliminating the need to copy and paste the exact file path into a web browser to view the file. 

Anyone is free to modify the script as they wish to add functionality, improve its logic or modify its behavior. The only thing I ask is that you leave a citation in comments somewhere at the top of the script indicating that I am the original creator of the script.